UniTimeBot/requirements.txt
streamlit
faiss-cpu
PyPDF2
sentence-transformers
python-dotenv